Sucessfully mad a Recorder object and a channel inser component that will be the template for drying up my control pad.
