+++ client/public/index.html
@@ -13,6 +13,6 @@
     <noscript>
       You need to enable JavaScript to run this app.
     </noscript>
-    <script src="../dist/bundle.js"></script>
+    <script src="/dist/bundle.js"></script>
   </body>
 </html>
